Fix review suggestion for refactoring the qr lib
