+++ src/ActiveLogin.Authentication.BankId.AspNetCore/Qr/IBankIdQrCodeGenerator.cs
@@ -1,4 +1,4 @@
-namespace ActiveLogin.Authentication.BankId.AspNetCore.Models
+namespace ActiveLogin.Authentication.BankId.AspNetCore.Qr
 {
     /// <summary>
     /// This interface is used to generate QR codes for BankID
