+++ test/ActiveLogin.Authentication.BankId.AspNetCore.Test/BankId_Authentication_Tests.cs
@@ -181,8 +181,7 @@ namespace ActiveLogin.Authentication.BankId.AspNetCore.Test
 	        // Arrange mocks
             var autoLaunchOptions = new BankIdLoginOptions(new List<string>(), null, false, true, false, false, String.Empty);
 	        var mockProtector =  new Mock<IBankIdLoginOptionsProtector>();
-            var mockQrCodeGenerator = new Mock<IBankIdQrCodeGenerator>();
-	        mockProtector
+            mockProtector
 		        .Setup(protector => protector.Unprotect(It.IsAny<string>()))
 		        .Returns(autoLaunchOptions);
 
@@ -199,7 +198,6 @@ namespace ActiveLogin.Authentication.BankId.AspNetCore.Test
 			        services =>
                     {
                         services.AddTransient(s => mockProtector.Object);
-                        services.AddTransient(s => mockQrCodeGenerator.Object);
 			        })
 		        .CreateClient();
 
@@ -237,7 +235,6 @@ namespace ActiveLogin.Authentication.BankId.AspNetCore.Test
             // Arrange mocks
             var autoLaunchOptions = new BankIdLoginOptions(new List<string>(), null, false, true, false, false, String.Empty);
             var mockProtector = new Mock<IBankIdLoginOptionsProtector>();
-            var mockQrCodeGenerator = new Mock<IBankIdQrCodeGenerator>();
             mockProtector
                 .Setup(protector => protector.Unprotect(It.IsAny<string>()))
                 .Returns(autoLaunchOptions);
@@ -256,7 +253,6 @@ namespace ActiveLogin.Authentication.BankId.AspNetCore.Test
                     services =>
                     {
                         services.AddTransient(s => mockProtector.Object);
-                        services.AddTransient(s => mockQrCodeGenerator.Object);
                         services.AddSingleton<IBankIdApiClient>(s => testBankIdApi);
                     })
                 .CreateClient();
